requirements.txt
spacy==3.7.2
accelerate
anthropic
anyio
confection==0.1.4
dask[dataframe]
einops
fastapi
fschat==0.2.36
google-generativeai
infini-gram
jinja2
jupyterlab
llvmlite
matplotlib
mistral-common
mistralai
nltk
numba
numpy
openai
pandas
protobuf
psutil
pyarrow
pydantic
pygments
pyyaml
ray
rich
scikit-learn
scipy
sentence-transformers
sentencepiece
shapely
tiktoken
tokenizers
torch==2.4.0
torchvision==0.19.0
tqdm
transformers
triton
typer
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3
vllm>=0.3.0
wasabi
weasel
websockets
xformers
datasketch
art  # for ArtPrompt
tenacity  # for ArtPrompt
boto3
bpe
wandb